More work on BitMaster. Resources and prototype.
Added more functionality and better layout to BitMaster, which is still
under development. Now it shows values (powers of two) under bits and
also has autofocus on inputs.

Added icons pack to the project (in case we keep the current icon).

Started prototyping AutoMister game to gain familiarity with finite
automata. Currently is a ChatGPT developed stub to exemplify the idea.
It is not intended to be used right now and future versions (if any)
may be substantially different.